requirements.txt
django>=6.0
folium>=0.20
python-dotenv>=1.0
pillow>=12.0
psycopg2-binary>=2.9
pydantic>=2.12
gunicorn>=25.0
requests>=2.32
